Edited dashboard time view to 1h - 3m.
Prometheus has a time delay in processing the data for 200+ containers.  It is unable to show values for all series at the current browser time, so this new dashboard plots from now-60m to now-3m which gives prometheus time to get the data imported and back out to grafana.